requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image/Video Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# 3D Mesh Processing & Visualization
open3d>=0.18.0
trimesh>=3.23.0
pyrender>=0.1.45
pyglet>=2.0.0

# Deep Learning Utilities
transformers>=4.35.0
timm>=0.9.0
einops>=0.7.0

# Math & Scientific Computing
scipy>=1.11.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1

# SAM-3D-Body (선택사항, 설치 방법 참조: SAM3D_INTEGRATION.md)
# sam-3d-body  # pip install sam-3d-body 또는 GitHub에서 설치

# Optional: ONNX Runtime (for inference optimization)
# onnx>=1.15.0
# onnxruntime-gpu>=1.16.0

# Optional: PyTorch3D (for advanced mesh operations)
# Note: May require manual installation via conda
# pytorch3d>=0.7.4
